Reorganize and rename frontend files
Also restructure Shape to more properly define ShapeMold
