+++ frontend/src/components/footer/Footer.tsx
