+++ frontend/src/components/input/SliderPrompt.tsx
