+++ frontend/src/components/output/SubmissionError.ts
