+++ frontend/src/components/stat/StatComponent.module.css
