+++ frontend/src/components/stat/chart/BarChart.tsx
