+++ frontend/src/components/stat/chart/DataLabelsOptions.ts
