+++ frontend/src/components/stat/chart/LineChart.tsx
