+++ frontend/src/shape/shapes/BarStatShape.ts
@@ -1,0 +1,5 @@
+import { StatShape } from "./StatShape";
+
+export type BarStatShape = StatShape & {
+    type: "bar";
+};
