+++ frontend/src/shape/shapes/LineStatShape.ts
@@ -1,0 +1,5 @@
+import { StatShape } from "./StatShape";
+
+export type LineStatShape = StatShape & {
+    type: "line";
+};
