+++ frontend/src/shape/shapes/StatShape.ts
@@ -1,0 +1,10 @@
+import { ArrayElement } from "../../magic/ArrayElement";
+
+export const StatShapeTypes = ["bar", "line"];
+export type StatRowsShape = {
+    [key: string]: number | number[];
+};
+export type StatShape = {
+    type: ArrayElement<typeof StatShapeTypes>;
+    rows: StatRowsShape;
+};
